combine activationresult.py to activation.py
